% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_checks.R
\name{ll_check_connection}
\alias{ll_check_connection}
\title{Check for internet connection.}
\usage{
ll_check_connection()
}
\value{
Stop if no connection detected
}
\description{
Check for internet connection.
}
\examples{
check_connection()
}
