% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_metadata.R
\name{ll_clean_authors}
\alias{ll_clean_authors}
\title{Clean up authors and title from ll_query-type output.}
\usage{
ll_clean_authors(x)
}
\arguments{
\item{x}{string, the whole response from ChatGPT}
}
\value{
a dataframe.
}
\description{
Clean up authors and title from ll_query-type output.
}
