% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ll_clean_journals}
\alias{ll_clean_journals}
\title{Clean up title, year, and journal from ll_query-type output.}
\usage{
ll_clean_journals(x)
}
\arguments{
\item{x}{string, the whole response from ChatGPT}
}
\value{
a dataframe.
}
\description{
Clean up title, year, and journal from ll_query-type output.
}
