% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_references.R
\name{ll_extract_references}
\alias{ll_extract_references}
\title{Mine ChatGPT for pdf metadata: works cited information.}
\usage{
ll_extract_references(
  file_path,
  model = "gpt-5.4-mini",
  api_key = Sys.getenv("OPENAI_API_KEY"),
  clean_references = TRUE
)
}
\arguments{
\item{file_path}{string, A single pdf file or directory containing pdfs.}

\item{model}{string, which OpenAI ChatGPT model to use.}

\item{api_key}{string, the API key if you want to supply it manually. There is no free one, but used with a minimal model is very cheap. https://openai.com/api/pricing/.}

\item{clean_references}{logical, whether to output a cleaned list of dataframes. If FALSE, returns raw ChatGPT output. This can be useful in case ChatGPT goes off script and hallucinates.}
}
\value{
a list of dataframe(s).
}
\description{
Mine ChatGPT for pdf metadata: works cited information.
}
