% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_metadata.R
\name{ll_find_institution_type}
\alias{ll_find_institution_type}
\title{Perform logical checks on a vector of institutions using ChatGPT.}
\usage{
ll_find_institution_type(
  text,
  model = "gpt-5.4-mini",
  api_key = Sys.getenv("OPENAI_API_KEY"),
  logical_check = "in the United States"
)
}
\arguments{
\item{text}{A vector of strings containing cleaned institutions. These might be the output of `ll_pull_institutions`.}

\item{model}{string, which OpenAI ChatGPT model to use.}

\item{api_key}{string, the API key if you want to supply it manually. There is no free one, but used with a minimal model is very cheap. https://openai.com/api/pricing/.}

\item{logical_check}{string, a phrase that can be answered "Yes" or "no" when combined with "Is this institution...". For example, "in the United States" or "a Minority Serving Institution".}
}
\value{
A data frame containing T/F data for each item in `text`.
}
\description{
For example, we might be interested in checking whether author institutions are a nonprofit, private company, or university.
}
\examples{
ll_find_institution_type("Johns Hopkins University", logical_check = "an institution with >5000 students")
}
