% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_metadata.R
\name{ll_journal_corrections}
\alias{ll_journal_corrections}
\title{Clean up / Remove some journal specific stuff.}
\usage{
ll_journal_corrections(x)
}
\arguments{
\item{x}{Collapsed output from pdftools that needs to be cleaned up}
}
\value{
Cleaned string.
}
\description{
Clean up / Remove some journal specific stuff.
}
