% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_metadata.R
\name{ll_pull_institutions}
\alias{ll_pull_institutions}
\title{Extract "Institutions" from a messy vector using ChatGPT.}
\usage{
ll_pull_institutions(
  text,
  model = "gpt-5.4-mini",
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{text}{A vector of strings containing messy data.}

\item{model}{string, which OpenAI ChatGPT model to use.}

\item{api_key}{string, the API key if you want to supply it manually. There is no free one, but used with a minimal model is very cheap. https://openai.com/api/pricing/.}
}
\value{
A vector containing clean data.
}
\description{
Extract "Institutions" from a messy vector using ChatGPT.
}
