% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ll_check_api_key}
\alias{ll_check_api_key}
\title{Check for the OpenAI API Key.}
\usage{
ll_check_api_key()
}
\value{
A message.
}
\description{
Check for the OpenAI API Key.
}
\examples{
ll_check_api_key() # Lets you know if an API key is correctly stored for next steps
}
