% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ll_check_file_structure}
\alias{ll_check_file_structure}
\title{Check if provided value is a file or a directory and return files.}
\usage{
ll_check_file_structure(file_path)
}
\arguments{
\item{file_path}{string, A single pdf file or directory containing pdfs.}
}
\value{
file name(s)
}
\description{
Check if provided value is a file or a directory and return files.
}
