% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ll_extract_journals}
\alias{ll_extract_journals}
\title{Mine ChatGPT for pdf metadata: Title, Year, and Journal.}
\usage{
ll_extract_journals(
  file_path,
  model = "gpt-4o-mini",
  api_key = Sys.getenv("OPENAI_API_KEY"),
  clean_journals = TRUE
)
}
\arguments{
\item{file_path}{string, A single pdf file or directory containing pdfs.}

\item{model}{string, which OpenAI ChatGPT model to use.}

\item{api_key}{string, the API key if you want to supply it manually. There is no free one, but used with a minimal model is very cheap. https://openai.com/api/pricing/.}

\item{clean_journals}{logical, whether to output a cleaned dataframe of journals. If FALSE, returns raw ChatGPT output. This can be useful in case ChatGPT goes off script and hallucinates.}
}
\value{
a dataframe.
}
\description{
Mine ChatGPT for pdf metadata: Title, Year, and Journal.
}
