% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ll_query}
\alias{ll_query}
\title{Perform a query against OpenAI's API.}
\usage{
ll_query(
  source,
  prompt,
  model = "gpt-4o-mini",
  api_key = Sys.getenv("OPENAI_API_KEY")
)
}
\arguments{
\item{source}{string, either a file or string of text that data will be mined
from. "pdf/journal_article.pdf" or "Here is a very long string of text I
would like to query directly"}

\item{prompt}{string, a prompt for ChatGPT.}

\item{model}{string, which OpenAI ChatGPT model to use.}

\item{api_key}{string, the API key if you want to supply it manually. There is no free one, but used with a minimal model is very cheap. https://openai.com/api/pricing/.}
}
\value{
Raw content from API
}
\description{
See https://openai.com/api/pricing/ for information on different models.
}
